started moving Dong Lee's code from visualodometry into features
